# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BeMore SME Access Initiative API
  description: |
    API for the BeMore live engagement and data capture platform targeting South African
    property developers, landowners, student accommodation operators, and built environment
    professionals seeking institutional funding partnerships through PBSA.

    ## Authentication
    Admin endpoints use **HttpOnly cookie-based JWT** authentication. On successful login,
    the server sets two cookies:
    - `bm_token` — HttpOnly JWT cookie (not accessible via JavaScript)
    - `bm_csrf` — CSRF token cookie (readable by client)

    State-changing admin requests (POST, PATCH, PUT, DELETE) require the CSRF token
    sent via the `X-CSRF-Token` header (double-submit cookie pattern).

    ## Rate Limiting
    All endpoints are rate-limited. Standard headers (`RateLimit-*`) are returned.
    | Tier | Window | Max Requests | Applies To |
    |------|--------|-------------|------------|
    | Health | 1 min | 200 | `GET /health` |
    | Public | 15 min | 100/IP | Application submit, lookup, POPIA, public settings |
    | Admin | 15 min | 300 | All admin endpoints |
    | Auth | 15 min | 10 | `POST /auth/login` |
    | Tracking | 15 min | 300/IP | `POST /track/*` |

    ## Application Status Flow
    `new` → `reviewing` → `shortlisted` → `invited` → `funded`

    ## Auto-Tagging
    Applications are automatically tagged on submission based on form data:
...

    ## POPIA Compliance
...
  version: 3.0.0
  contact:
...
servers:
...
paths:
...
  /applications:
    post:
      tags: [Applications]
      summary: Submit application (public)
      description: |
        Submit a new application. Duplicate prevention is enforced — same email + userType
        combination returns 409 with the existing reference number. Phone numbers are
        normalized to +27 format. Auto-tagging is applied based on form data.
        Rate limited to 100 requests per 15 minutes per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationInput'
      responses:
        '201':
          description: Application submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Application'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Duplicate application (same email + userType)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: An application with this email and category already exists
                  refNumber:
                    type: string
                    example: BM-09C5B0EF
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      tags: [Applications]
      summary: List applications (admin)
      description: Paginated, filterable list of all applications. Rate limited to 300 req/15min.
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, reviewing, shortlisted, invited, funded, all]
        - name: userType
          in: query
          schema:
            type: string
            enum: [developer, landowner, investor, student, professional, aspiring, all]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Paginated list of applications
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Application'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /applications/upload:
    post:
      tags: [Applications]
      summary: Upload an attachment (public, multipart)
      description: |
        Spec §8.1 — multipart/form-data upload for the Built Environment
        Professional CV (Q13). Accepts a single file in the `file` field. No
        authentication required, but a valid `X-CSRF-Token` header is mandatory
        (CSRF double-submit). Rate limited to 100 req/15min per IP.

        - Max file size: **5 MB** (`MAX_FILE_BYTES` in `uploadService.js`).
        - Accepted MIME types: `application/pdf`, `application/msword`,
          `application/vnd.openxmlformats-officedocument.wordprocessingml.document`.
        - Files are persisted to `${UPLOAD_DIR}/cv/<uuid>.<ext>` (Railway
          mounts a persistent volume at `/app/uploads`).
        - The returned `storedAs` value must be passed back in
          `attachments[].storedAs` on `POST /applications` to link the file
          to the submission.
      security: []
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF or Word document, max 5 MB.
      responses:
        '200':
          description: Upload accepted and persisted to disk
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  filename:
                    type: string
                    description: Sanitised display filename (`[a-zA-Z0-9 ._-]`).
                    example: cv-thabo-mokoena.pdf
                  storedAs:
                    type: string
                    description: UUID-based filename on disk; pass to `POST /applications` via `attachments[].storedAs`.
                    example: 7c8a3a2e-1234-4abc-9def-0123456789ab.pdf
                  size:
                    type: integer
                    description: File size in bytes.
                    example: 124356
                  mimeType:
                    type: string
                    example: application/pdf
        '400':
          description: |
            Upload rejected. `code` is one of:
            - `NO_FILE` — no `file` field in the request.
            - `FILE_TOO_LARGE` — file exceeds 5 MB.
            - `INVALID_MIME_TYPE` — not a PDF or Word document.
            - `UPLOAD_ERROR` — other multer-level failure.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  code:
                    type: string
                    enum: [NO_FILE, FILE_TOO_LARGE, INVALID_MIME_TYPE, UPLOAD_ERROR]
                  message:
                    type: string
                    example: File exceeds 5 MB limit
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          description: Disk-write failure (`STORAGE_ERROR`).
  /applications/lookup:
    post:
      tags: [Applications]
      summary: Status lookup (public)
      description: |
        Allows applicants to check their application status using reference number and email.
        No authentication required. Rate limited to 100 req/15min per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refNumber, email]
              properties:
                refNumber:
                  type: string
                  example: BM-09C5B0EF
                email:
                  type: string
                  format: email
                  example: thabo@example.co.za
      responses:
        '200':
          description: Application status found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      refNumber:
                        type: string
                        example: BM-09C5B0EF
                      firstName:
                        type: string
                        example: Thabo
                      userType:
                        type: string
                        enum: [developer, landowner, investor, student, professional, aspiring]
                      status:
                        type: string
                        enum: [new, reviewing, shortlisted, invited, funded]
                      tags:
                        type: array
                        items:
                          type: string
                      summitAccess:
                        type: boolean
                      submittedAt:
                        type: string
                        format: date-time
                      updatedAt:
                        type: string
                        format: date-time
        '404':
          description: No matching application found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: No application found. Please check your reference number and email.
  /applications/stats:
    get:
      tags: [Applications]
      summary: Application statistics (admin)
      description: Aggregated statistics for the admin dashboard.
      responses:
        '200':
          description: Aggregated statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /applications/export/csv:
    get:
      tags: [Applications]
      summary: Export applications as CSV (admin)
      description: |
        Downloads all applications as a CSV file. Formula injection prevention
        is applied (cells starting with `=`, `+`, `-`, `@` are prefixed).
      responses:
        '200':
          description: CSV file download
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="bemore-applications-2026-04-24.csv"
          content:
            text/csv:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /applications/bulk-status:
    post:
      tags: [Applications]
      summary: Bulk status update (admin)
      description: |
        Update status for up to 100 applications at once. Triggers email notifications
        for shortlisted, invited, and funded status changes. Requires CSRF token.
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids, status]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    description: MongoDB ObjectId
                  minItems: 1
                  maxItems: 100
                status:
                  type: string
                  enum: [new, reviewing, shortlisted, invited, funded]
      responses:
        '200':
          description: Bulk update successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      updated:
                        type: integer
                        example: 15
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /applications/send-reminders:
    post:
      tags: [Applications]
      summary: Send summit reminder emails (admin)
      description: Send summit reminder emails to up to 100 applicants. Requires CSRF token.
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    description: MongoDB ObjectId
                  minItems: 1
                  maxItems: 100
      responses:
        '200':
          description: Reminders sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      sent:
                        type: integer
                        example: 25
        '401':
          $ref: '#/components/responses/Unauthorized'
  /applications/{refNumber}/attachment/{storedAs}/signed:
    get:
      tags: [Applications]
      summary: Download attachment via signed link (public, no auth)
      description: |
        Spec §8.5 — public download of an attachment using an HMAC-signed URL.
        No authentication and no CSRF required; integrity is enforced by the
        `expires` + `sig` query parameters.

        Signature scheme (`backend/src/utils/signedLinks.js`):
        - `sig = HMAC-SHA256(JWT_SECRET, "${refNumber}|${storedAs}|${expires}")` (hex)
        - `expires = floor(Date.now()/1000) + 300` (default TTL 5 minutes,
          `SIGNED_LINK_TTL_SECONDS`)

        Signed links are typically embedded in the `attachments[].downloadUrl`
        field returned by `POST /applications/data-export`.

        Response is the raw file bytes with
        `Content-Disposition: attachment; filename="<original-filename>"`.
      security: []
      parameters:
        - name: refNumber
          in: path
          required: true
          schema:
            type: string
            pattern: '^BM-[A-Z0-9]{4,12}$'
          example: BM-09C5B0EF
        - name: storedAs
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-zA-Z0-9._-]+$'
          example: 7c8a3a2e-1234-4abc-9def-0123456789ab.pdf
        - name: expires
          in: query
          required: true
          description: Unix epoch (seconds) at which the link expires.
          schema:
            type: integer
            example: 1747005600
        - name: sig
          in: query
          required: true
          description: Hex HMAC-SHA256 signature (see description).
          schema:
            type: string
            pattern: '^[a-f0-9]{64}$'
      responses:
        '200':
          description: File bytes streamed with attachment disposition.
          headers:
            Content-Disposition:
              schema:
                type: string
                example: attachment; filename="cv-thabo-mokoena.pdf"
            Content-Type:
              schema:
                type: string
                example: application/pdf
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          description: '`BAD_SIGNATURE` — HMAC mismatch or malformed `sig`/`expires`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: '`APPLICATION_NOT_FOUND` or `ATTACHMENT_NOT_FOUND`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: |
            Link no longer usable. `code` is one of:
            - `LINK_EXPIRED` — `expires` is in the past.
            - `FILE_MISSING_ON_DISK` — DB record exists but the file was removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: bm_token
      description: HttpOnly JWT cookie set by POST /auth/login
...
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
  responses:
    Unauthorized:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            error:
              code: AUTH_TOKEN_INVALID
              message: Invalid or expired authentication token
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                - field: email
                  message: Valid email required
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: Too many requests. Please try again later.
security:
  - CookieAuth: []